refactor: extract shared PermissionsPanel from onboarding step
The permission list (status rows, grant/request actions, misplaced-bundle
warning) moves to components/widget/permissions/permissions-panel.tsx so it
can be reused outside the onboarding wizard. PermissionsStep keeps only the
onboarding chrome (StepShell, eyebrow, copy).
